o1vm: rename riscv32i in riscv32im
In order to match the newly added M instruction type
