+++ o1vm/src/interpreters/mod.rs
@@ -4,7 +4,7 @@ pub mod keccak;
 /// An interpreter for the MIPS instruction set.
 pub mod mips;
 
-/// An interpreter for the RISC-V 32I instruction set, following the specification
+/// An interpreter for the RISC-V 32IM instruction set, following the specification
 /// on
 /// [riscv.org](https://riscv.org/wp-content/uploads/2019/12/riscv-spec-20191213.pdf).
-pub mod riscv32i;
+pub mod riscv32im;
